Move statemachine settings
