+++ addons/medical_statemachine/config.cpp
@@ -14,5 +14,6 @@ class CfgPatches {
     };
 };
 
+#include "ACE_Settings.hpp"
 #include "Statemachine.hpp"
 #include "CfgEventHandlers.hpp"
