+++ addons/medical_statemachine/stringtable.xml
@@ -1,0 +1,66 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project name="ACE">
+    <Package name="Medical_Statemachine">
+        <Container name="Settings">
+            <Key ID="STR_ACE_Medical_statemachine_fatalInjuryCondition_DisplayName">
+                <English>Fatal Injury Player</English>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_fatalInjuryCondition_Description">
+                <English>Defines when Player can receive a fatal injury</English>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_fatalInjuryConditionAI_DisplayName">
+                <English>Instant death AI</English>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_fatalInjuryConditionAI_Description">
+                <English>Defines if AI will be able to die instantly</English>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_unconsciousConditionAI_DisplayName">
+                <English>Forbid unconscious AI</English>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_unconsciousConditionAI_Description">
+                <English>Defines if AI will be denied to go unconscious and will die instead</English>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_fatalInjuryCondition_InCardiacArrest">
+                <English>In Cardiac Arrest</English>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_cardiacArrestTime_DisplayName">
+                <English>Cardiac Arrest time (seconds)</English>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_cardiacArrestTime_Description">
+                <English>Defines how long it takes to die from cardiac arrest?</English>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_enableUnconsciousnessAI_DisplayName">
+                <English>AI Unconsciousness</English>
+                <Russian>Потеря сознания ботами</Russian>
+                <Polish>Nieprzytomność AI</Polish>
+                <Spanish>Inconsciencia IA</Spanish>
+                <German>KI-Bewusstlosigkeit</German>
+                <Czech>Bezvědomí AI</Czech>
+                <Portuguese>Inconsciência da IA</Portuguese>
+                <French>Inconscience des IA</French>
+                <Hungarian>AI eszméletlenség</Hungarian>
+                <Italian>Incoscienza IA</Italian>
+                <Japanese>AI の気絶</Japanese>
+                <Korean>인공지능 기절</Korean>
+                <Chinesesimp>AI无意识</Chinesesimp>
+                <Chinese>AI無意識</Chinese>
+            </Key>
+            <Key ID="STR_ACE_Medical_statemachine_enableUnconsciousnessAI_Description">
+                <English>Allow AI to go unconscious</English>
+                <Russian>Позволить ботам терять сознание</Russian>
+                <Polish>Czy AI może być nieprzytomne od odniesionych obrażeń?</Polish>
+                <Spanish>Permita a la IA caer inconsciente</Spanish>
+                <German>KI kann bewusstlos werden</German>
+                <Czech>Umožňuje AI upadnout do bezvědomí</Czech>
+                <Portuguese>Permite IA ficar inconsciente</Portuguese>
+                <French>Autoriser les IA à tomber inconscients</French>
+                <Hungarian>Engedélyezi az AI eszméletének elvesztését</Hungarian>
+                <Italian>Permetti alle IA di diventare incoscienti</Italian>
+                <Japanese>AI が気絶をするように</Japanese>
+                <Korean>인공지능도 기절에 빠지게 합니다</Korean>
+                <Chinesesimp>允许AI进入无意识状态</Chinesesimp>
+                <Chinese>允許AI進入無意識狀態</Chinese>
+            </Key>
+        </Container>
+    </Package>
+</Project>
